Refactor a bunch
Coming to the conclusion that the user code needs to manage the per
frame buffers and more.. the render function now takes a bunch of
arguments.

I've tried to create a frame abstraction that holds all the resources
for a frame in flight / element of the swapchain.. but there are certain
things that make this not nice. E.g., you have to index
image_acquired_semaphores differently from the rest and the swapchain
elements need to be recreated on resize..